fix(operator-ui): restore HTTP default and dedupe WS panels
